all comment except productPaging, try-catch in getBy DAO

+++ src/main/java/sb/utils/ListToPageConverter.java
@@ -10,7 +10,7 @@ public class ListToPageConverter<T> {
 
     public Page<T> toPage(List<T> tList, Pageable pageable) {
         int start = (int) pageable.getOffset();
-        int end =   (int) (Math.min((start + pageable.getPageSize()), tList.size()));
+        int end = Math.min((start + pageable.getPageSize()), tList.size());
 
         return new PageImpl<T>(tList.subList(start, end), pageable, tList.size());
     }
